fix(toolbar.js): Set the URL param even if it is equal to the default value (so that the filter set in the cookie is updated)
Before: if you had "Remember Category Pagination" option turned on, filters are memoized in browser cookie.

However, if the user wanted to change the filters back to their default value (e.g. from limit 15 to 30 and then back to limit 15), they wouldn't get applied - because the URL param would get skipped (as it would be equal to the default value), and thus navigating to it didn't change anything.

This commit fixes it.